refactor: extract circuit breaker into circuit_breaker.py, deduplicate wrapper (#161, #164)
- New naas/library/circuit_breaker.py: RedisCircuitBreakerStorage,
  _get_circuit_breaker, and with_circuit_breaker helper
- with_circuit_breaker replaces the duplicated ~30-line try/except
  block in netmiko_send_command and netmiko_send_config
- Redis client is now lazily initialised on first call to
  _get_circuit_breaker, preventing import-time failure when Redis
  is unavailable at worker startup
- decode_responses removed (aligns with other Redis clients)
- netmiko_lib.py reduced from 346 to ~170 lines
