+++ changes/164.bugfix.md
@@ -1,0 +1,1 @@
+Fix module-level Redis client in netmiko_lib.py initialised at import time; now lazily initialised on first use in circuit_breaker.py
